Move the `Issue`, `Issues` and `IssueState` types into the issues module.
