+++ src/gitlab.rs
@@ -327,7 +327,7 @@ impl GitLab {
     ///
     /// Because we need to search (and thus query the GitLab server possibly multiple times), this
     /// _can_ be a slow operation if there is many issues in the project.
-    pub fn get_issue(&self, namespace: &str, name: &str, iid: i64) -> Result<::Issue> {
+    pub fn get_issue(&self, namespace: &str, name: &str, iid: i64) -> Result<::issues::Issue> {
         // We first need to find the specific project.
         let project = self.get_project(namespace, name)
             .chain_err(|| format!("cannot get project '{}/{}'", namespace, name))?;
@@ -335,7 +335,7 @@ impl GitLab {
         // Closure to search for the item, possibly returning multiple match on multiple pages.
         let query_gitlab_closure = || self.issues().project(project.id);
         // Closure to find the right item in the found list on the page.
-        let iter_find_closure = |ref issue: &::Issue| issue.iid == iid;
+        let iter_find_closure = |ref issue: &::issues::Issue| issue.iid == iid;
 
         self.get_paginated_from_project(query_gitlab_closure, iter_find_closure)
     }
