+++ src/serde_types.in.rs
@@ -38,17 +38,6 @@ pub enum ListingVisibility {
 
 
 #[derive(Debug, Serialize, Deserialize)]
-pub enum IssueState {
-    #[serde(rename = "opened")]
-    Opened,
-    #[serde(rename = "closed")]
-    Closed,
-    #[serde(rename = "reopened")]
-    Reopened,
-}
-
-
-#[derive(Debug, Serialize, Deserialize)]
 pub enum UserState {
     #[serde(rename = "active")]
     Active,
@@ -244,30 +233,3 @@ pub struct User {
     pub avatar_url: Option<String>,
     pub web_url: Option<String>
 }
-
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Issue {
-    pub id: i64,
-    pub iid: i64,
-    pub project_id: i64,
-    pub title: String,
-    pub description: String,
-    pub state: IssueState,
-    pub created_at: String,  // FIXME: Use date type?
-    pub updated_at: String,  // FIXME: Use date type?
-    pub labels: Vec<String>,
-    pub milestone: Option<Milestone>,
-    pub assignee: Option<User>,
-    pub author: User,
-    pub subscribed: bool,
-    pub user_notes_count: i64,
-    pub upvotes: i64,
-    pub downvotes: i64,
-    pub due_date: Option<String>,  // FIXME: Use date type?
-    pub confidential: bool,
-    pub web_url: Option<String>
-}
-
-
-pub type Issues = Vec<Issue>;
